github: update PR template to focus on "why"
Previously, the PR template was very focused on describing the changes
that were made, leaving less space for a presentation of why the given
change is even needed in the first place.

Although it did incentivize the user to provide an issue (where previous
discussion would've been held), it's not always the case that a given PR
is a 1:1 with the issue that it references.

By forcing ourselves to reiterate why we're submitting the PR, it might
prompt us to better think about whether we're solving the problem in the
right way.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>