atc, tf: WIP update user activity on every action
I have written a failing testflight test because now the output of
`fly active-users` is too long. I suspect that the root cause is that the
accessor handler is updating user activity for unauthenticated requests and
also on some kind of system activity (like the TSA or workers?). If you build
this version (don't forget to `go install ./fly`), you will see the surprising
three lines of output from `fly active-users`.

concourse/concourse#5119

Signed-off-by: Jamie Klassen <cklassen@pivotal.io>