atc: remove scope and duration from lidar metrics
following the rationale that for an operator, individual scopes might
not be useful, and that the actual status of those checks is more
important.

being just counters, it should be simpler for an operator to reason
about the values being emitted for these metrics (how many checks
occurred, what their status were, how many have been enqueued, etc).

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>